refactor: encapsulate identifier resolution in BasicValueIdCreator
- Move identifier property resolution and sequence detection logic from SimpleIdBinder to BasicValueIdCreator.
- Simplify SimpleIdBinder.bindSimpleId by using the new encapsulated methods.
- Update relevant unit tests to verify the refactored logic.
